updated models, added controllers, updated SQL scripts

+++ src/main/java/com/interactiveresume/Interactive/Resume/Backend/jpa/auth/RefreshTokenJPARepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RefreshTokenJPARepository extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenJPARepository extends JpaRepository<RefreshToken, Long> {
 
     Optional<RefreshToken> findByToken(String token);
